refactor: 문구 수정 및 car 테스트를 다른 파일로 분리
